Pull in kivymd master changes (#2)
* Added check to see if any widget of the subclass snackbar is present. If so don't show new one.

* Fixed the state of the `md_bg_color` property of the `MDCard` class which did not retain its value when changing the application color scheme

* Removed `StackFloatingButton` rule from `registry`

* Fix typo

* Possibly a typo in the class name

Doc example for `SmartTileWithLabel` was using `SmartTileWithStar`. Fixed that. The example is working as expected.

* Added snackbar_x, snackbar_y and snackbar_animation_dir

* Fixed PR as per comments

* Fixed PR as per comments

* Made anchor_title an OptionProperty and updated doc

* Run pre-commit

* Fixed DropdownMenu item responsive area shifting

This is a fix for #552. The idea is to use different list classes for when there's only text, only a left icon with text, only a right content with text, or when both left icon and right content is present because with the current implementation it needs to remove a widget when there's no left icon which causes removal of `touchable widgets` from the MDlist class used and so the touch cannot propagate which results in shifting of responsive area

* Fixed comment formatting

* Fixed comment formatting

* Fixed Doc for Base List tertiary_text_color

* Fix https://github.com/kivymd/KivyMD/issues/564

* Added radius_right to backdrop

* Fixed example to match picture

* Fixed example

* Fixed changed property comments

* Fixed example again

* Edit doc string

Co-authored-by: Dalton Woodside-Moffatt <dwoodsidemoffatt@gmail.com>
Co-authored-by: macbookair <kivydevelopment@gmail.com>
Co-authored-by: Shashi Ranjan <shashiranjankv@gmail.com>
Co-authored-by: KivyMD Bot <69076719+KivyMD-Bot@users.noreply.github.com>